requirements.txt
streamlit
arxiv
wikipedia
ddgs
pandas
plotly
networkx
requests
beautifulsoup4
google-genai
python-dotenv
pypdf
xhtml2pdf
markdown
kaleido
cairosvg
langgraph
pydantic-ai[google]
mcp
langsmith
nest_asyncio
